+++ dev/admin/js/config.js
@@ -1,0 +1,7 @@
+// set up some the config for RequireJS
+requirejs.config({
+    baseUrl: 'js',
+    paths: {
+        jquery: '../../js/jquery'
+    }
+});
