+++ dev/admin/js/script.js
@@ -1,784 +1,6 @@
-window.onload = init;
 
-function init() {
-  dataTypes();
-  var elements = document.getElementsByTagName('body');
-  var page = elements[0].id;
-
-  switch(page) {
-    case 'works':
-      // display the old works
-      displayWorks();
-      // cancel the studio in the textarea
-      var cancelWork = document.getElementById('cancelWork');
-      cancelWork.onclick = clear;
-      // filter the old works
-      filterWorks();
-      // delete old works after clicking on the button
-      deleteBadStuff('works');
-      // grab the new studio
-      var saveWork = document.getElementById('saveWork');
-      saveWork.onclick = addWork;
-
-      var saveWork = document.getElementById('saveWork');
-      var workImage = document.getElementById('workImage');
-      workImage.addEventListener('change', function() {
-        saveWork.removeAttribute('disabled');
-        saveWork.onclick = gatherData;
-      }, false);
-
-      break;
-    case 'studio':
-      // display all the old studio photos
-      displayStudio();
-      // cancel the studio in the textarea
-      var cancelStudio = document.getElementById('cancelStudio');
-      cancelStudio.onclick = clear;
-      deleteBadStuff('studio');
-      // grab the new studio
-      var saveStudio = document.getElementById('saveStudio');
-      var studioImage = document.getElementById('studioImage');
-      studioImage.addEventListener('change', function() {
-        saveStudio.removeAttribute('disabled');
-        saveStudio.onclick = gatherData;
-      }, false);
-      break;
-    case 'news':
-      // display the old news
-      displayNews();
-      // cancel the news in the textarea
-      var cancelNews = document.getElementById('cancelNews');
-      cancelNews.onclick = clear;
-      // delete old news after clicking on the button
-      deleteBadStuff('news');
-      // grab the new news
-      var saveNews = document.getElementById('saveNews');
-      var newsContent = document.getElementById('newsContent');
-      newsContent.addEventListener('input', function() {
-        saveNews.removeAttribute('disabled');
-        saveNews.onclick = gatherData;
-      }, false);
-      break;
-    case 'cv':
-      $.get('../resume-raw.html', function(data) {
-        $('.resume .resume-content').html(data);
-      });
-      break;
-    case 'contact':
-      break;
-  }
-}
-
-
-
-
-
-// we need this for one thing…
-function capitalizeFirstLetter(string) {
-  return string.charAt(0).toUpperCase() + string.slice(1);
-}
-
-
-
-
-
-
-// create new work
-function Work(id, title, year, media, description, dimension_d, dimension_w, dimension_h, available, images, date) {//, images) {
-  this.id = id,
-  this.title = title,
-  this.year = year,
-  this.media = media,
-  this.description = description,
-  this.dimension_d = dimension_d,
-  this.dimension_w = dimension_w,
-  this.dimension_h = dimension_h,
-  this.available = available,
-  // this.image = image//,
-  this.images = images,
-  this.date = date
-}
-// create new studio
-function Studio(id, image, date) {
-	this.id = id,
-	this.image = image,
-	this.date = date
-}
-// create new news
-function News(id, description, date) {
-	this.id = id,
-	this.description = description,
-	this.date = date
-}
-
-
-
-
-
-// delete the old studio & news
-function deleteOld() {
-  $('.delete').click(function() {
-    // give this a var
-    var deleteBtn = $(this);
-    var dataId = $(this.parentElement).data('id');
-    console.log(dataId);
-    var section = $('body').attr('id');
-    deleteData(section, dataId);
-  });
-}
-// some arrays we will need as we go along
-var sectionArr = ['design', 'drawing', 'furnishings', 'painting', 'sculpture', 'students', 'studio', 'news'];
-// loop through the sections and get all the data available
-function dataTypes() {
-	for(var i = 0; i < sectionArr.length; i++) {
-		// lets get the data and store it locally
-		getData(sectionArr[i]);
-	}
-}
-// GET ALL THE DATA
-function getData(section) {
-	var request = new XMLHttpRequest();
-	request.open('GET', '../data/' + section + '.json');
-
-	request.onreadystatechange = function() {
-		var div = document.getElementById('results');
-		if(this.readyState == this.DONE && this.status == 200) {
-			var type = request.getResponseHeader('Content-Type');
-			if(this.responseText != null) {
-				var json = JSON.parse(this.responseText);
-				var keys = Object.keys(json);
-				localStorage.setItem(section, JSON.stringify(json));
-			}
-		}
-	}
-	request.send();
-}
-function reloadData(section) {
-  localStorage.removeItem(section);
-  getData(section);
-  switch(section) {
-    case 'studio':
-      clear(section);
-      setTimeout(function(){
-        var oldStudio = document.getElementById('oldStudio');
-        oldStudio.innerHTML = '';
-        displayStudio();
-      },300);
-      break;
-    case 'news':
-      clear(section);
-      setTimeout(function(){
-        var oldNews = document.getElementById('oldNews');
-        oldNews.innerHTML = '';
-        displayNews();
-      },300);
-      break;
-    case 'furnishings':
-    case 'sculpture':
-    case 'drawing':
-    case 'painting':
-    case 'design':
-    case 'students':
-      clear(section);
-      setTimeout(function(){
-        var oldWorks = document.getElementById('oldWorks');
-        oldWorks.innerHTML = '';
-        displayWorks();
-      },300);
-      break;
-  }
-}
-// CLEAR THE DATA ON THE PAGE
-function clear() {
-  var section = $('body').attr('id');
-  switch(section) {
-    case 'studio':
-      var saveStudio = document.getElementById('saveStudio');
-      saveStudio.setAttribute('disabled', 'disabled');
-
-      var uploadProcess = document.getElementById('uploadProcess');
-      uploadProcess.setAttribute('style', 'width: 0%;');
-
-      var studioErr = $('.msg.error');
-      $(studioErr).text('');
-      document.getElementById('studioImage').value = '';
-      break;
-    case 'news':
-      var saveNews = document.getElementById('saveNews');
-      saveNews.setAttribute('disabled', 'disabled');
-
-      var newsErr = $('.msg.error');
-      $(newsErr).text('');
-      document.getElementById('newsContent').value = '';
-      break;
-    case 'works':
-      var saveWork = document.getElementById('saveWork');
-      saveWork.setAttribute('disabled', 'disabled');
-
-      var uploadProcess = document.getElementById('uploadProcess');
-      uploadProcess.setAttribute('style', 'width: 0%;');
-
-      var workErr = $('.msg.error');
-      $(workErr).text('');
-      document.getElementById('workImage').value = '';
-      document.getElementById('newTitle').value = '';
-      document.getElementById('newYear').value = '';
-      document.getElementById('newMedia').value = '';
-      document.getElementById('newDescription').value = '';
-      document.getElementById('newDimension_d').value = '';
-      document.getElementById('newDimension_w').value = '';
-      document.getElementById('newDimension_h').value = '';
-      break;
-  }
-}
-
-
-
-
-function uploadWorkImg(today, timeId, fd, file) {
-  var stringWork = addWork(today, timeId, fd, file);
-  var fileName = file.name;
-  var fileArr = fileName.split('.');
-  var imgIndex = fileArr.length - 1;
-  var imgSuff = fileArr[imgIndex];
-  var xhr = new XMLHttpRequest();
-  xhr.open('POST', 'functions/uploadwork.php', true);
-
-  xhr.upload.onprogress = function(e) {
-    if (e.lengthComputable) {
-      var percentComplete = (e.loaded / e.total) * 100;
-      var uploadProcess = document.getElementById('uploadProcess');
-      uploadProcess.setAttribute('style', 'width: ' + percentComplete + '%;');
-      if(percentComplete === 100) {
-        // do something… or do nothing!
-      }
-    }
-  };
-
-  xhr.onload = function() {
-    if (this.status == 200) {
-      saveStuff('studio', stringWork);
-    };
-  };
-  xhr.send(fd);
-}
-function addWork(today, id, fd, file) {
-  var fileName = file.name;
-  var fileArr = fileName.split('.');
-  var imgIndex = fileArr.length - 1;
-  var imgSuff = fileArr[imgIndex];
-  // grab the news content
-  var workImage = document.getElementById('workImage').value;
-  // get the error div
-  var workErr = $('.msg.error');
-  // check to see if the
-  if(workImage === '') {
-    // display the error
-    $(workErr).text('please add a work image...');
-  } else {
-    // clear the error when new is entrered
-    $(workErr).text('');
-    // build out the news
-    var newWork = new Work(id, title, year, media, description, dimension_w, dimension_h, available, images, date);
-    // stringify the news
-  	var stringWork = JSON.stringify(newWork);
-    // do something with the news
-    return stringWork;
-  }
-}
-function getWorkSection() {
-  // get the category filters
-  var addWorkCategory = $('#addWorkCategory');
-  // get the children of the list
-  var addWorkItems = addWorkCategory[0].children;
-  // loop through all the items in the filter list
-  for(var i = 0; i < addWorkItems.length; i++) {
-    // make sure that the radio button is checked
-    if( addWorkItems[i].children[0].checked === true ) {
-      // return the section
-      return newSection = addWorkItems[i].children[0].id;
-    }
-  }
-}
-function filterWorks() {
-  // get all of the work filters
-  // NEED TO ADD IN SOME BUSINESS ABOUT ADDING/REMOVIONG CHECKS
-  $('#filterWorks .list-item label').click(function() {
-    // get the attribute of each label
-    var section = $(this).attr('for');
-    // remove any hiding going on
-    $('#oldWorks').children('.works').show();
-    // add hiding to the siblings of the type clicked on
-    $('#oldWorks').children('.' + section).siblings('.works:not(' + '.' + section + ')').hide();
-  });
-}
-// lets swap the main image with related ones
-function swapWorks() {
-  $('.studio-img-thmb').click(function() {
-    $(this).parent('.list-item').siblings().children('.studio-img-thmb').removeClass('active');
-    $(this).addClass('active');
-    // get this and it's parent
-    var imgId = $(this).attr('data-imgid');
-    var parent = $(this).closest('.module-section');
-    var id = $(parent).attr('id');
-    var mainImg = $(parent).children('.studio-img');
-    // parts of the image
-    var imgPath = '../img/works/';
-    var imgNum = id;
-    var imgSize = '_l-';
-    var imgSuff = '.jpg';
-    var imgIndex = imgId;
-    // build the image
-    $(mainImg).attr('src', imgPath + imgNum + imgSize + imgIndex + imgSuff );
-  });
-}
-
-
-
-
-// DO SOME DATE STUFF
-function showDate(data) {
-  var date = new Date(data.date);
-  var day = date.getDate().toString();
-  var monthNum = date.getMonth();
-  var monthNames = ['JAN', 'FEB', 'MAR', 'APR', 'MAY', 'JUN', 'JUL', 'AUG', 'SEP', 'OCT', 'NOV', 'DEC'];
-  var month = monthNames[monthNum];
-  var year = (date.getFullYear()).toString();
-  var year = year.slice(-2);
-  var studioDate = day + month + year;
-  return studioDate;
-}
-// START SAVING STUFF
-function saveStuff(section, data) {
-  var msg = document.getElementById('messaging');
-  $.ajax({
-      type: 'GET',
-      url: 'functions/saveit.php?section=' + section + '&data=' + encodeURIComponent(data),
-      dataType: 'JSON',
-      success: function(ret){
-        msg.classList.add('success');
-        msg.innerHTML = 'Your ' + section + ' has been added!';
-        msg.innerHTML = ret;
-      }
-  });
-  reloadData(section);
-}
-// DELETE STUFF
-function deleteStuff(section, index) {
-  var msg = document.getElementById('messaging');
-  $.ajax({
-      type: 'GET',
-      url: 'functions/delete.php?section=' + section + '&index=' + encodeURIComponent(index),
-      dataType: 'JSON',
-      success: function(ret){
-        msg.classList.add('success');
-        msg.innerHTML = 'Your ' + section + ' has been deleted!';
-        msg.innerHTML = ret;
-      }
-  });
-  setTimeout(function(){
-    reloadData(section);
-  },300);
-}
-// START DELETING OLD STUFF
-function deleteBadStuff(section) {
-  var sectionCap = capitalizeFirstLetter(section);
-  $('#old' + sectionCap).on('click', '.delete', function() {
-    var deleteBtn = $(this);
-    if( section === 'works' ) {
-      var deleteParent = deleteBtn.parents('.module-section');
-    } else {
-      var deleteParent = deleteBtn[0].parentNode;
-    }
-    var dataId = $(deleteParent).attr('id');
-    var dataIndex = $(deleteParent).attr('data-id');
-    document.getElementById(dataId).remove();
-    var dataSection = $(this).parents('.module-section').data('section');
-    if( section === 'works' ) {
-      deleteStuff(dataSection, dataIndex);
-    } else {
-      deleteStuff(section, dataIndex);
-    }
-  });
-}
-
-
-
-
-
-function uploadStudioImg(today, timeId, fd, file) {
-  var stringStudio = addStudio(today, timeId, fd, file);
-  var fileName = file.name;
-  var fileArr = fileName.split('.');
-  var imgIndex = fileArr.length - 1;
-  var imgSuff = fileArr[imgIndex];
-  var xhr = new XMLHttpRequest();
-  xhr.open('POST', 'functions/upload.php', true);
-
-  xhr.upload.onprogress = function(e) {
-    if (e.lengthComputable) {
-      var percentComplete = (e.loaded / e.total) * 100;
-      var uploadProcess = document.getElementById('uploadProcess');
-      uploadProcess.setAttribute('style', 'width: ' + percentComplete + '%;');
-      if(percentComplete === 100) {
-        // do something… or do nothing!
-      }
-    }
-  };
-
-  xhr.onload = function() {
-    if (this.status == 200) {
-      saveStuff('studio', stringStudio);
-    };
-  };
-  xhr.send(fd);
-}
-function addStudio(today, id, fd, file) {
-  var fileName = file.name;
-  var fileArr = fileName.split('.');
-  var imgIndex = fileArr.length - 1;
-  var imgSuff = fileArr[imgIndex];
-  // grab the news content
-  var studioImage = document.getElementById('studioImage').value;
-  // get the error div
-  var studioErr = $('.msg.error');
-  // check to see if the
-  if(studioImage === '') {
-    // display the error
-    $(studioErr).text('please add a studio shot...');
-  } else {
-    // clear the error when new is entrered
-    $(studioErr).text('');
-    // build out the news
-    var newStudio = new Studio(id, '../img/studio/' + id + '_l.' + imgSuff, today);
-    // stringify the news
-  	var stringStudio = JSON.stringify(newStudio);
-    // do something with the news
-    return stringStudio;
-  }
-}
-
-
-
-
-
-// DISPLAY THE WORKS
-var workNums = [];
-var workSum = [];
-var workAll = [];
-function displayWorks() {
-  // loop through the sectionArr
-  for(var i = 0; i < sectionArr.length - 2; i++) {
-    // add to the workAll arr
-    workAll.push(sectionArr[i]);
-  }
-  for(var i = 0; i < workAll.length; i++) {
-    var allWorks = workAll[i];
-    var allWorks = JSON.parse( localStorage.getItem( allWorks ) );
-    for( var key in allWorks ) {
-      // get the length of each section
-      dbNum( allWorks[key].length );
-      for( var j = 0; j < allWorks[key].length; j++ ) {
-        var oldWorks = document.getElementById('oldWorks');
-        // create the main div
-        var div = document.createElement('div');
-        div.setAttribute('class', 'module-section work works ' + key);
-        div.setAttribute('id', allWorks[key][j].id);
-        div.setAttribute('data-section', key);
-        div.setAttribute('data-id', j);
-        // create the image
-        var img = document.createElement('img');
-        img.setAttribute('class', 'studio-img');
-        // build the image
-        var imgPath = '../img/works/';
-        var imgNum = allWorks[key][j].id;
-        var imgSize = '_l-';
-        var imgSuff = '.jpg';
-        img.setAttribute('src', imgPath + imgNum + imgSize + 0 + imgSuff);
-        // create the date span
-        var span = document.createElement('span');
-        span.setAttribute('class', 'date');
-        span.innerHTML = allWorks[key][j].date;
-        // create the title heading
-        var heading1 = document.createElement('h3');
-        heading1.setAttribute('class', 'heading');
-        heading1.innerHTML = 'Title';
-        // create the title heading
-        var input1 = document.createElement('input');
-        input1.setAttribute('class', 'form-input');
-        input1.setAttribute('type', 'text');
-        input1.setAttribute('value', allWorks[key][j].title);
-        // create the year heading
-        var heading2 = document.createElement('h3');
-        heading2.setAttribute('class', 'heading');
-        heading2.innerHTML = 'Year';
-        // create the year heading
-        var input2 = document.createElement('input');
-        input2.setAttribute('class', 'form-input');
-        input2.setAttribute('type', 'text');
-        input2.setAttribute('value', allWorks[key][j].year);
-        // create the media heading
-        var heading3 = document.createElement('h3');
-        heading3.setAttribute('class', 'heading');
-        heading3.innerHTML = 'Media';
-        // create the media heading
-        var input3 = document.createElement('input');
-        input3.setAttribute('class', 'form-input');
-        input3.setAttribute('type', 'text');
-        input3.setAttribute('value', allWorks[key][j].media);
-        // create the description heading
-        var heading4 = document.createElement('h3');
-        heading4.setAttribute('class', 'heading');
-        heading4.innerHTML = 'Description';
-        // create the description heading
-        var input4 = document.createElement('textarea');
-        input4.setAttribute('class', 'form-input');
-        input4.setAttribute('type', 'text');
-        input4.innerHTML = allWorks[key][j].description;
-        // create the dimensions heading
-        var heading4 = document.createElement('h3');
-        heading4.setAttribute('class', 'heading');
-        heading4.innerHTML = 'Dimensions';
-        // create the dimensions heading
-        var input4 = document.createElement('input');
-        input4.setAttribute('class', 'form-input dimension');
-        input4.setAttribute('type', 'text');
-        input4.setAttribute('value', allWorks[key][j].dimension_d);
-        // create the dimensions heading
-        var input5 = document.createElement('input');
-        input5.setAttribute('class', 'form-input dimension');
-        input5.setAttribute('type', 'text');
-        input5.setAttribute('value', allWorks[key][j].dimension_w);
-        // create the dimensions heading
-        var input6 = document.createElement('input');
-        input6.setAttribute('class', 'form-input dimension');
-        input6.setAttribute('type', 'text');
-        input6.setAttribute('value', allWorks[key][j].dimension_h);
-        // create the available heading
-        var heading5 = document.createElement('h3');
-        heading5.setAttribute('class', 'heading');
-        heading5.innerHTML = 'Available';
-        // create the available heading
-        var input7 = document.createElement('input');
-        input7.setAttribute('class', 'form-input');
-        input7.setAttribute('type', 'text');
-        input7.setAttribute('value', allWorks[key][j].available);
-        // create the related images heading
-        var heading6 = document.createElement('h3');
-        heading6.setAttribute('class', 'heading');
-        heading6.innerHTML = 'All Images';
-        // create the related images heading
-        var ulRelated = document.createElement('ul');
-        ulRelated.setAttribute('class', 'list inline related');
-        // loop through it all
-        for( var m = 0; m <= allWorks[key][j].images - 1; m++ ) {
-          var liRelated = document.createElement('li');
-          liRelated.setAttribute('class', 'list-item');
-          // img element stuff
-          var allImgs = document.createElement('img');
-          allImgs.setAttribute('class', 'studio-img-thmb');
-          if( m === 0 ) {
-            allImgs.setAttribute('class', 'studio-img-thmb active');
-          }
-          // build the image bits
-          var imgPath = '../img/works/';
-          var imgNum = allWorks[key][j].id;
-          var imgSize = '_m-';
-          var imgSuff = '.jpg';
-          var imgIndex = m;
-          // build the image
-          allImgs.setAttribute('src', imgPath + imgNum + imgSize + imgIndex + imgSuff);
-          allImgs.setAttribute('data-imgid', imgIndex);
-          // append it all
-          liRelated.appendChild(allImgs);
-          ulRelated.appendChild(liRelated);
-        }
-        // save module
-        var saveDiv = document.createElement('div');
-        saveDiv.setAttribute('class', 'module-save');
-        // create the save button
-        var button1 = document.createElement('button');
-        button1.setAttribute('class', 'save button');
-        button1.setAttribute('disabled', 'disabled');
-        button1.setAttribute('type', 'submit');
-        button1.setAttribute('name', 'edit');
-        button1.setAttribute('value', 'save');
-        button1.innerHTML = 'Save';
-        // create the delete button
-        var button2 = document.createElement('button');
-        button2.setAttribute('class', 'delete button');
-        button2.setAttribute('type', 'submit');
-        button2.setAttribute('name', 'edit');
-        button2.setAttribute('value', 'delete');
-        button2.innerHTML = 'Delete';
-        // append everything
-        div.appendChild(img);
-        div.appendChild(heading1);
-        div.appendChild(input1);
-        div.appendChild(heading2);
-        div.appendChild(input2);
-        div.appendChild(heading3);
-        div.appendChild(input3);
-        div.appendChild(heading4);
-        div.appendChild(input4);
-        div.appendChild(heading4);
-        div.appendChild(input4);
-        div.appendChild(input5);
-        div.appendChild(input6);
-        div.appendChild(heading5);
-        div.appendChild(input7);
-        div.appendChild(heading6);
-        div.appendChild(ulRelated);
-        // insert bigtime
-        div.insertBefore(span, div.firstChild);
-        saveDiv.appendChild(button1);
-        saveDiv.appendChild(button2);
-        div.appendChild(saveDiv);
-        oldWorks.appendChild(div);
-      }
-    }
-  }
-  // lets swap the main image with related ones
-  swapWorks();
-}
-// DISPLAY THE STUDIO
-function displayStudio() {
-  var data = JSON.parse( localStorage.getItem('studio') );
-  for (var key in data) {
-    var data = data[key].reverse();
-    // get the length of each section
-    dbNum( data.length );
-    for(var i = 0; i < data.length; i++) {
-      // old studio stuff
-      var oldStudio = document.getElementById('oldStudio');
-      var div = document.createElement('div');
-      div.setAttribute('class', 'module-section work studio');
-      // console.log( data );
-      div.setAttribute('id', data[i].id);
-      div.setAttribute('data-id', (data.length - 1) - i);
-      // img stuff
-      var img = document.createElement('img');
-      img.setAttribute('class', 'studio-img');
-      img.setAttribute('src', data[i].image);
-      // date stuff
-      var studioDate = showDate(data[i]);
-      var span = document.createElement('span');
-  		span.setAttribute('class', 'date');
-  		span.innerHTML = studioDate;
-      // button stuff
-      var button = document.createElement('button');
-      button.setAttribute('class', 'delete button');
-      button.setAttribute('type', 'submit');
-      button.setAttribute('name', 'edit');
-      button.setAttribute('value', 'delete');
-      button.innerHTML = 'Delete';
-      // append it all
-      div.appendChild(img);
-      div.insertBefore(span, div.firstChild);
-      div.appendChild(button);
-      oldStudio.appendChild(div);
-    }
-  }
-}
-// DISPLAY THE NEWS
-function displayNews() {
-  var newsData = JSON.parse( localStorage.getItem('news') );
-  for (var key in newsData) {
-    var news = newsData[key].reverse();
-    // get the length of each section
-    dbNum( news.length );
-    for(var i = 0; i < news.length; i++) {
-      // old news stuff
-      var oldNews = document.getElementById('oldNews');
-      var div = document.createElement('div');
-      div.setAttribute('class', 'modules-section news-item');
-      div.setAttribute('id', news[i].id);
-      div.setAttribute('data-id', (news.length - 1) - i);
-      div.innerHTML = news[i].description;
-      // date stuff
-      var newsDate = showDate(news[i]);
-      var span = document.createElement('span');
-      span.setAttribute('class', 'news-item-date');
-      span.innerHTML = newsDate;
-      // button stuff
-      var button = document.createElement('button');
-      button.setAttribute('class', 'delete button');
-      button.setAttribute('type', 'submit');
-      button.setAttribute('name', 'edit');
-      button.setAttribute('value', 'delete');
-      button.innerHTML = 'Delete';
-      // append it all
-      div.appendChild(button);
-      div.insertBefore(span, div.firstChild);
-      oldNews.appendChild(div);
-    }
-  }
-}
-
-
-function gatherData() {
-  var section = $('body').attr('id');
-  switch(section) {
-    case 'works':
-      var workImg = document.getElementById('workImage');
-      if( workImg.files[0] ) {
-          var today = new Date();
-          var timeId = today.getTime();
-          var file = workImg.files[0];
-          var fd = new FormData();
-          fd.append('workImage', file);
-          fd.append('id', timeId);
-          uploadWorkImg(today, timeId, fd, file);
-      }
-      break;
-    case 'studio':
-      var studioImg = document.getElementById('studioImage');
-      if( studioImg.files[0] ) {
-          var today = new Date();
-          var timeId = today.getTime();
-          var file = studioImg.files[0];
-          var fd = new FormData();
-          fd.append('studioImage', file);
-          fd.append('id', timeId);
-          uploadStudioImg(today, timeId, fd, file);
-      }
-      break;
-    case 'news':
-      // get today's date
-      var today = new Date();
-      // create an id
-      var id = today.getTime();
-      // grab the news content
-      var newsContent = document.getElementById('newsContent').value;
-      // get the error div
-      var newsErr = $('.msg.error');
-      // check to see if the
-      if(newsContent === '') {
-        // display the error
-        $(newsErr).text('please add some news...');
-      } else {
-        // clear the error when new is entrered
-        $(newsErr).text('');
-        // build out the news
-        var newNews = new News(id, newsContent, today);
-        // stringify the news
-      	var stringNews = JSON.stringify(newNews);
-        // do something with the news
-        saveStuff('news', stringNews);
-      }
-      break;
-  }
-}
-
-
-
-
-
-
-// GET NUMBER OF ITEMS IN DB
-newNum = 0;
-function dbNum(num) {
-    var dbNum = document.getElementById('dbNum');
-    newNum += num;
-    dbNum.innerHTML = newNum;
-}
+//Load common code that includes config, then load the app logic for this page.
+requirejs(['config'], function (config) {
+    requirejs(['jquery']);
+    requirejs(['init']);
+});
